requirements.txt
flask
python-dotenv
pymongo
hume
pytest
pylint
chardet
hume[stream] 